# fix: Use correct C++ without UB (data race), that produce same or faster code