Another argument is `annotate_ignore_thread_sanitizer_guard` should be used only to mark false positive issues, not the real one.

Benchmark results:
AMD Ryzen 9 9950X, bare metal
```
============================================================================
[...]elox/exec/benchmarks/AtomicsBench.cpp     relative  time/iter   iters/s
============================================================================
std_atomic_bool_write_seq_cst                              1.16min    14.36m
std_atomic_bool_write_release                                4.50s   222.17m
std_atomic_bool_write_relaxed                                4.48s   223.00m
one_way_flag_write                                         55.20ms     18.12
one_way_flag_unsafe_write                                  55.57ms     17.99
std_atomic_bool_read_seq_cst                               29.66ms     33.72
std_atomic_bool_read_acquire                               30.10ms     33.22
std_atomic_bool_read_relaxed                               30.00ms     33.33
one_way_flag_read                                          29.84ms     33.51
one_way_flag_unsafe_read                                   59.68ms     16.76
```